% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RS.R
\name{ConnectRS}
\alias{ConnectRS}
\title{Connect to AWS Redshift DBMS using JDBC Driver}
\usage{
ConnectRS(yrurl = NULL, host = NULL, dbname = NULL, user = NULL,
  driverClass = NULL, classPath = NULL, identifier.quote = NULL)
}
\arguments{
\item{yrurl}{A character string of the URL to the cluster of the DBMS.
Paste only the following part of the URL to function input:
yrurl = "<JDBCURL>:<port>/<dbname>?user=<user>&". If yrurl is supplied,
ConnectRS requires no further arguments as inputs.}

\item{host}{A character string of the host}

\item{dbname}{A character string of the DB name}

\item{user}{A character string of the username}

\item{driverClass}{A character string of with the Class of the JDBC Driver.
Check
\href{https://docs.aws.amazon.com/redshift/latest/mgmt/configure-jdbc-connection.html#obtain-jdbc-url}{AWS Documentation}
for up to date driver class and adjust following link if necessary
"com.amazon.redshift.jdbc42.Driver".
If new driver class 43 is available enter e.g.:
"com.amazon.redshift.jdbc43.Driver"}

\item{classPath}{A character string of the path to the folder containing
the JDBC-Driver (a .jar file)}

\item{identifier.quote}{Optional argument; NA by default; can impact
functions as e.g.
\code{\link[DBI]{dbWriteTable}} or \code{\link[DBI]{dbReadTable}}
if not chosen correctly; see also \code{\link[RJDBC]{JDBC}} or
\href{https://docs.aws.amazon.com/redshift/latest/dg/r_names.html}{AWS Documentation}}
}
\value{
Returns a connection object which can be used to query
the AWS Redshift DBMS with e.g.: \code{\link[DBI]{dbGetQuery}}.
}
\description{
Function to establish JDBC driven connection to AWS Redshift DBMS.
Credentials and cluster information required. Communication with
Redshift DB, which runs on a modified PostgreSQL dialect, can
require special
\href{https://docs.aws.amazon.com/redshift/latest/dg/c_SQL_commands.html}{Redshift PostgreSQL code}.
}
\details{
Either provide: yrurl OR all other arguments.
Connection to DB not visible in RStudio connection pane. Disconnect
by using e.g. \link[DBI]{dbDisconnect}

For details on the JDBC cluster URL see:
\href{https://docs.aws.amazon.com/redshift/latest/mgmt/configure-jdbc-connection.html#obtain-jdbc-url}{AWS Documentation}
}
\examples{
\dontrun{
Not generally executable examples:

con <- ConnectRS(yrurl = "jdbc:redshift://redshift.adsbexchange.com:5439/adsbx?user=random.user&")
(when the url is given)

where:
<JDBCURL> <- jdbc:redshift://redshift.adsbexchange.com
<port> <- 5439
<dbname> <- adsbx
user=<user> <- random.user
Password retrieved via popo-up

dbDisconnect(con)

con <- ConnectRS(host = "redshift.adsbexchange.com", dbname = "adsbx",
user = "random.user", driverClass = "com.amazon.redshift.jdbc42.Driver",
classPath = "C:/.../JDBC/RedshiftJDBC42-1.2.15.1025.jar", identifier.quote = "`")
Password retrieved via popo-up

dbDisconnect(con)}

}
