% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RS_dplyr.R
\name{ConnectRSdplyr}
\alias{ConnectRSdplyr}
\title{Setup Connection for querying AWS Redshift with dplyr}
\usage{
ConnectRSdplyr(host, user, dbname, port = 5439)
}
\arguments{
\item{host}{A character string for the DB host}

\item{user}{A character string for the Username}

\item{dbname}{A character string for the Name of the DB}

\item{port}{Port is set to 5439 (class(port): "numeric") by default.
This is the default TCP port for AWS Redshift DBMS.}
}
\value{
Returns a "Formal Class PQConnection" object which can be
used to e.g. query the AWS Redshift DBMS with help of the
\code{\link[dplyr]{dplyr}}, more specifically the
\code{\link[dbplyr]{dbplyr}} package.
}
\description{
Function to establish PostgreSQL driven connection to AWS Redshift DBMS.
Communication with DB via this channel requires only dplyr verbs, no
PostgreSQL code is required.
}
\details{
Generally not executable examples:
Connecting to PostgreSQL AWS Redshift database hosted by
"redshift.adsbexchange.com" for user "erikseumegoettingen",
with database name "adsbx". Enter password in pop-up.

con <- ConnectRSdplyr(host = 'redshift.adsbexchange.com',
user = 'erikseumegoettingen', dbname = 'adsbx')

dbDisconnect(con)
dbIsValid(con) # confirms that connection is closed by returning FALSE
}
