% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historic_adsb_exchange.R
\name{GetHistoricADSBExchange}
\alias{GetHistoricADSBExchange}
\title{Query position tracking data from the ADSB-Exchange Amazon Redshift cluster}
\usage{
GetHistoricADSBExchange(con, ica, start, end, limit = 10)
}
\arguments{
\item{con}{A connection object which is obtained using \code{\link{ConnectRS}}}

\item{ica}{A character string of the ICAO code of an aircraft.
To find a specific ICAO code of an aircraft, go e.g. to
\href{https://www.flightradar24.com/59.09,13.09/6}{Flightradar24.com} and
choose an aircraft. The left hand pop-up window contains a section
describing the aircraft type, registration and MODE-S CODE which is the ICAO
code.}

\item{start}{A character string of the start date and time when the aircraft
should start being tracked. Format: "YYYY-MM-DD hh:mm:ss" Time zone must be
UTC}

\item{end}{A character string of the end date and time when the aircraft
should stop being tracked. Format: "YYYY-MM-DD hh:mm:ss" Time zone must be
UTC}

\item{limit}{Is the maximum number of rows queried from the DB. Is set to
10 by default to prevent large queries.}
}
\value{
Returns a data.frame which can be used for
analysis with the plotting function \code{\link{PlotAirways}}
}
\description{
Queries the ADSB-Exchange Amazon Redshift Cluster for the parameters:
ICAO code; 13 digit UNIX (milliseconds timestamp) at UTC;
latitude; longitude; to obtain a data.frame as input for further analysis.
Every observation of the aircraft position (latitude, longitude)
is linked with a timestamp at every point. This date is itslef linked to
aircraft identifier.
}
\details{
Function arguments ica, start and end are character strings.
Enter start and end in UTC date/time. Local times must be coverted
beforehand into UTC Time by using e.g. \code{\link{anytime}}

GetHistoricADSBExchange(con, ica, start, end, limit = 10)
}
\examples{
\dontrun{Not generally executable example:
GetHistoricADSBExchange(con = con, ica = "86D660", start = "2018-08-31 01:00:00",
end = "2018-08-31 12:10:00", limit = 10)}

}
