% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_data.R
\name{GetLiveData}
\alias{GetLiveData}
\title{GetLiveData}
\usage{
GetLiveData(point = "", radius = 100, country = "", type = "")
}
\arguments{
\item{point}{A vector containing the latitude and longitude of a specific point or
a string containing the name of a city to determine a specicfic position as inital point.}

\item{radius}{The number of kilometers which should be considered as radius
for a circular selection area around the point}

\item{country}{A string containing the country where an aircraft is registered}

\item{type}{A string containing a specific aircraft type}
}
\value{
A dataframe containing latitude, longitude, operator, type, country, start and destination
as well as the icao (identification) of the selected aircrafts
}
\description{
This function provides information about current flights,
like position, destination and operator in a dataframe format.
Flights can be selected by different parameters.
}
\examples{
flights_berlin <- GetLiveData(point = c(52.5243, 13.4063))
flights_berlin2 <- GetLiveData(point = "Berlin")
flights_vienna50 <- GetLiveData(point = "Vienna", radius = 50)

flights_australia <- GetLiveData(country = "Australia")

flights_B738 <- GetLiveData(type = "B738")

flights_mixture <- GetLiveData(point = "Paris", radius = 500, country = "Germany", type = "A320")

}
\references{
For a table of the selectable cities see \code{\link[maps:world.cities]{world.cities}}.
For further information see also \url{https://www.adsbexchange.com/data/}
}
