% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NoPlanes}
\alias{NoPlanes}
\title{NoPlanes}
\usage{
NoPlanes(point = "", radius = 100)
}
\arguments{
\item{point}{A string containing the name of a city.}

\item{radius}{The number of kilometers which should be considered as radius
for a circular area around the point.}
}
\value{
A sentence which defines the area and the number of aircraft in it.
"There are currently \emph{x} airplanes in the area \emph{y} km around \emph{city}."
}
\description{
This function returns the number of aircrafts in a selected area.
}
\details{
See also \code{\link{GetLiveData}}
}
\examples{
NoPlanes("New York")
NoPlanes("Copenhagen", radius = 200)

}
