% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotAirways}
\alias{PlotAirways}
\title{PlotAirways}
\usage{
PlotAirways(con, ica, start, end, limit = 10)
}
\arguments{
\item{con}{A connection object which is obtained using \code{\link{ConnectRS}}}

\item{ica}{A character string of the ICAO code of an aircraft (MODE-S CODE).
To find a specific MODE-S CODE of an aircraft, go e.g. to
\href{https://www.flightradar24.com/59.09,13.09/6}{Flightradar24.com} and
choose an aircraft. The left hand pop-up window contains a section
describing the aircraft type, registration and MODE-S CODE.}

\item{start}{A character string of the start date and time when the aircraft
should start being tracked. Format: "YYYY-MM-DD hh:mm:ss" Time zone must be
UTC}

\item{end}{A character string of the end date and time when the aircraft
should stop being tracked. Format: "YYYY-MM-DD hh:mm:ss" Time zone must be
UTC}

\item{limit}{Is the maximum number of rows queried from the DB. Is set to
10 by default to prevent large queries.}
}
\value{
A map plotting the past postion marks of an aircraft in a
selected time frame.
}
\description{
A function which plots the flightroute of an aircraft.
}
\details{
See also \code{\link{GetHistoricADSBExchange}}
for further information and how to get the con object
}
\examples{
\dontrun{
PlotAirways(con = con, ica = "4010EB", start = "2016-06-09 05:32:49",
end = "2016-06-09 18:55:35", limit = 3)}

}
