% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotPlanes}
\alias{PlotPlanes}
\title{PlotPlanes}
\usage{
PlotPlanes(point = "", radius = 100, country = "", type = "")
}
\arguments{
\item{point}{A vector containing the latitude and longitude of a specific point or
a string containing the name of a city to determine a specicfic position as inital point.}

\item{radius}{The number of kilometers which should be considered as radius
for a circular selection area around the point.}

\item{country}{A string containing the country where an aircraft is registred.}

\item{type}{A string containing a specific aircraft type.}
}
\value{
An interactive worldmap on which the position of the aircraft is displayed.
}
\description{
Creates an interactive map with markers at the current position of aircraft.
Clicking on the marker will provide further information about the aircraft.
}
\details{
For further examples see \code{\link{GetLiveData}}
}
\examples{
PlotPlanes("London", 50)
PlotPlanes(country = "Algeria")
PlotPlanes(type = "A319")

}
